model-trainer/requirements.txt
mlflow>=2.14.0
pyyaml>=6.0.1
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
